refactor: LottoProfitCalculator 를 LottoProfitCalculatorService 로 변경
LottoProfitCalculator 도메인으로 두었으나, 로또 당첨에 대한 수익률을 계산하는 기능으로, "도메인"이 아닌 "서비스"다.
