+++ src/main/java/lotto/Application.java
@@ -2,10 +2,10 @@ package lotto;
 
 import lotto.controller.LottoController;
 import lotto.domain.generator.LottoAutoGenerator;
-import lotto.domain.profit.LottoProfitCalculator;
 import lotto.domain.winning.LottoWinningSetPicker;
 import lotto.domain.winning.LottoWinningRankingCalculator;
 import lotto.domain.generator.LottoManualGenerator;
+import lotto.service.LottoProfitCalculatorService;
 import lotto.service.LottoStoreService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -19,9 +19,9 @@ public class Application {
         LottoWinningSetPicker lottoWinningSetPicker = new LottoWinningSetPicker();
         LottoStoreService lottoStoreService = new LottoStoreService(lottoAutoGenerator, lottoManualGenerator);
         LottoWinningRankingCalculator lottoWinningRankingCalculator = new LottoWinningRankingCalculator();
-        LottoProfitCalculator lottoProfitCalculator = new LottoProfitCalculator();
+        LottoProfitCalculatorService lottoProfitCalculatorService = new LottoProfitCalculatorService();
 
-        LottoController lottoController = new LottoController(inputView, outputView, lottoStoreService, lottoWinningSetPicker, lottoWinningRankingCalculator, lottoProfitCalculator);
+        LottoController lottoController = new LottoController(inputView, outputView, lottoStoreService, lottoWinningSetPicker, lottoWinningRankingCalculator, lottoProfitCalculatorService);
         lottoController.run();
     }
 }
